fix(auth): harden rate-control custody
